Rename nearest_points
