+++ cpp/include/cuspatial/nearest_points.cuh
