+++ cpp/tests/spatial/nearest_points/point_linestring_nearest_points_test.cu
@@ -19,7 +19,7 @@
 #include <cuspatial/error.hpp>
 #include <cuspatial/geometry/vec_2d.hpp>
 #include <cuspatial/iterator_factory.cuh>
-#include <cuspatial/point_linestring_nearest_points.cuh>
+#include <cuspatial/nearest_points.cuh>
 
 #include <rmm/device_vector.hpp>
